refactor(BlackJackResultService): 필요없는 서비스 삭제
