--- src/main/java/service/BlackJackResultService.java
@@ -1,5 +1,0 @@
-package service;
-
-public class BlackJackResultService {
-
-}
